requirements.txt
paramiko==3.1.0
pykeepass==4.0.3
random-password-generator==2.2.0
python-dotenv==1.0.0